requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.0.8
django-3-jet==1.0.8
django-celery-results==1.2.1
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
oauthlib==3.1.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
